+++ handlers/nasiya.py
@@ -1,29 +1,33 @@
+import requests
+import os
 from aiogram import Router, Bot
 from aiogram.types import Message, FSInputFile
 from aiogram.filters import Command
-from services.nasiya_pdf import generate_pdf
-import os
+from services.nasiya_pdf import nasiya_pdf
+from services.get_telegram import get_telegram_id
 
 router = Router()
 
-# 1. Komandaga javob beruvchi funksiya
+# 1. Foydalanuvchi komandasi bilan yuboriladigan versiya
 @router.message(Command("n_hisobot"))
 async def send_nasiya(message: Message, bot: Bot):
-    pdf_path = generate_pdf(logo_path='insell.png')
-    document = FSInputFile(path=pdf_path, filename=os.path.basename(pdf_path))
+    await send_nasiya_silent(bot, message.chat.id)
 
-    await bot.send_document(chat_id=message.chat.id, document=document)
+# Barcha foydalanuvchilarga mahsulot hisobotini yuborish (silent)
+async def send_nasiya_silent(bot: Bot, telegram_id: int):
+    try:
+        pdf_path = nasiya_pdf(telegram_id=telegram_id, logo_path='insell.png')
 
-    if os.path.exists(pdf_path):
-        os.remove(pdf_path)
+        if not os.path.exists(pdf_path):
+            print(f"❌ {telegram_id} uchun fayl topilmadi.")
+            return
 
-# 2. Silent versiya — cron uchun
-async def send_nasiya_silent(bot: Bot):
-    chat_id = 6951699115
-    pdf_path = generate_pdf(logo_path='insell.png')
-    document = FSInputFile(path=pdf_path, filename=os.path.basename(pdf_path))
+        document = FSInputFile(path=pdf_path, filename=os.path.basename(pdf_path))
+        await bot.send_document(chat_id=telegram_id, document=document)
 
-    await bot.send_document(chat_id=chat_id, document=document)
+        if os.path.exists(pdf_path):
+            os.remove(pdf_path)
 
-    if os.path.exists(pdf_path):
-        os.remove(pdf_path)
+    except Exception as e:
+        print(f"❌ {telegram_id} uchun nasiya hisobotini yuborishda xato: {e}")
+
